Make schema private
