+++ tembo-operator/src/extensions/toggle.rs
@@ -84,7 +84,7 @@ async fn toggle_extensions(
     Ok(ext_status_updates)
 }
 
-fn determine_updated_extensions_status(
+pub fn determine_updated_extensions_status(
     cdb: &CoreDB,
     all_actually_installed_extensions: Vec<ExtensionStatus>,
 ) -> Vec<ExtensionStatus> {
@@ -179,7 +179,7 @@ fn determine_updated_extensions_status(
     ext_status_updates
 }
 
-fn determine_extension_locations_to_toggle(cdb: &CoreDB) -> Vec<Extension> {
+pub fn determine_extension_locations_to_toggle(cdb: &CoreDB) -> Vec<Extension> {
     let mut extensions_to_toggle: Vec<Extension> = vec![];
     for desired_extension in &cdb.spec.extensions {
         let mut needs_toggle = false;
@@ -217,13 +217,7 @@ fn determine_extension_locations_to_toggle(cdb: &CoreDB) -> Vec<Extension> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{
-        apis::coredb_types::{CoreDB, CoreDBSpec, CoreDBStatus},
-        extensions::types::{
-            get_location_spec, get_location_status, Extension, ExtensionInstallLocation,
-            ExtensionInstallLocationStatus, ExtensionStatus,
-        },
-    };
+    use crate::apis::coredb_types::{CoreDB, CoreDBSpec};
 
     #[test]
     fn test_determine_updated_extensions_status_empty() {
@@ -238,361 +232,5 @@ mod tests {
         let all_actually_installed_extensions = vec![]; // modify as needed
         let result = determine_updated_extensions_status(&cdb, all_actually_installed_extensions);
         assert_eq!(result, vec![]);
-    }
-
-    #[test]
-    fn test_toggle_logic() {
-        let desired_extensions = vec![
-            Extension {
-                name: "ext3".to_string(),
-                description: None,
-                locations: vec![ExtensionInstallLocation {
-                    enabled: true,
-                    schema: None,
-                    database: "db1".to_string(),
-                    version: None,
-                }],
-            },
-            Extension {
-                name: "ext1".to_string(),
-                description: None,
-                locations: vec![
-                // Requesting to enable a currently disabled extension
-                ExtensionInstallLocation {
-                        enabled: true,
-                        schema: None,
-                        database: "db_where_its_available_and_disabled".to_string(),
-                        version: None,
-                },
-                // Requesting to disable a currently enabled extension
-                ExtensionInstallLocation {
-                    enabled: false,
-                    schema: None,
-                    database: "db_where_its_available_and_enabled".to_string(),
-                    version: None,
-                },
-                // Requesting to enable a currently disabled extension that is not currently in status
-                ExtensionInstallLocation {
-                    enabled: true,
-                    schema: None,
-                    database: "db_where_its_available_and_disabled_missing_from_status".to_string(),
-                    version: None,
-                },
-                // Requesting to disable a currently enabled extension that is not currently in status
-                ExtensionInstallLocation {
-                    enabled: false,
-                    schema: None,
-                    database: "db_where_its_available_and_enabled_missing_from_status".to_string(),
-                    version: None,
-                },
-                // This situation is if we toggled an extension to True, but it failed to enable
-                // And now we toggle it back to false
-                ExtensionInstallLocation {
-                    enabled: false,
-                    schema: None,
-                    database: "db_where_it_is_currently_in_error_having_tried_to_enable_and_failed".to_string(),
-                    version: None,
-                },
-                // This situation is if we toggled an extension to True, but it failed to enable
-                // because it wasn't installed, now we toggle it back to false
-                ExtensionInstallLocation {
-                    enabled: false,
-                    schema: None,
-                    database: "db_where_it_is_currently_in_error_having_tried_to_enable_and_failed_because_missing".to_string(),
-                    version: None,
-                },
-                // Requesting to enable an extension that is not installed
-                ExtensionInstallLocation {
-                    enabled: true,
-                    schema: None,
-                    database: "db_where_its_not_available".to_string(),
-                    version: None,
-                },
-                // Requesting to enable an extension that previously failed to enable
-                ExtensionInstallLocation {
-                    enabled: true,
-                    schema: None,
-                    database: "db_where_enable_failed".to_string(),
-                    version: None,
-                }
-                ],
-            },
-            Extension {
-                name: "ext2".to_string(),
-                description: None,
-                locations: vec![ExtensionInstallLocation {
-                    enabled: false,
-                    schema: None,
-                    database: "db1".to_string(),
-                    version: None,
-                }],
-            },
-        ];
-
-        let current_status = vec![ExtensionStatus {
-            name: "ext1".to_string(),
-            description: None,
-            locations: vec![
-                // Requesting to enable a currently disabled extension
-                ExtensionInstallLocationStatus {
-                    enabled: Some(false),
-                    database: "db_where_its_available_and_disabled".to_string(),
-                    schema: Some("public".to_string()),
-                    version: None,
-                    error: Some(false),
-                    error_message: None,
-                },
-                // Requesting to disable a currently enabled extension
-                ExtensionInstallLocationStatus {
-                    enabled: Some(true),
-                    database: "db_where_its_available_and_enabled".to_string(),
-                    schema: Some("public".to_string()),
-                    version: None,
-                    error: Some(false),
-                    error_message: None,
-                },
-                ExtensionInstallLocationStatus {
-                    enabled: Some(false),
-                    database: "db_where_it_is_currently_in_error_having_tried_to_enable_and_failed"
-                        .to_string(),
-                    schema: Some("public".to_string()),
-                    version: None,
-                    error: Some(true),
-                    error_message: Some("Failed to enable extension".to_string()),
-                },
-                ExtensionInstallLocationStatus {
-                    enabled: None,
-                    database:
-                        "db_where_it_is_currently_in_error_having_tried_to_enable_and_failed_because_missing"
-                            .to_string(),
-                    schema: Some("public".to_string()),
-                    version: None,
-                    error: Some(true),
-                    error_message: Some("Extension is not installed".to_string()),
-                },
-                ExtensionInstallLocationStatus {
-                    enabled: Some(false),
-                    database: "db_where_enable_failed".to_string(),
-                    schema: Some("public".to_string()),
-                    version: None,
-                    error: Some(true),
-                    error_message: Some("Failed to enable extension".to_string()),
-                },
-            ],
-        }];
-
-        let cdb = CoreDB {
-            metadata: Default::default(),
-            spec: CoreDBSpec {
-                extensions: desired_extensions,
-                ..Default::default()
-            },
-            status: Some(CoreDBStatus {
-                extensions: Some(current_status),
-                ..CoreDBStatus::default()
-            }),
-        };
-
-        let all_actually_installed_extensions = vec![
-            ExtensionStatus {
-                name: "ext1".to_string(),
-                description: None,
-                locations: vec![
-                    ExtensionInstallLocationStatus {
-                        enabled: Some(false),
-                        error: None,
-                        database: "db_where_its_available_and_disabled".to_string(),
-                        schema: Some("public".to_string()),
-                        version: None,
-                        error_message: None,
-                    },
-                    ExtensionInstallLocationStatus {
-                        enabled: Some(true),
-                        error: None,
-                        database: "db_where_its_available_and_enabled".to_string(),
-                        schema: Some("public".to_string()),
-                        version: None,
-                        error_message: None,
-                    },
-                    ExtensionInstallLocationStatus {
-                        enabled: Some(false),
-                        error: None,
-                        database: "db_where_its_available_and_disabled_missing_from_status".to_string(),
-                        schema: Some("public".to_string()),
-                        version: None,
-                        error_message: None,
-                    },
-                    ExtensionInstallLocationStatus {
-                        enabled: Some(true),
-                        error: None,
-                        database: "db_where_its_available_and_enabled_missing_from_status".to_string(),
-                        schema: Some("public".to_string()),
-                        version: None,
-                        error_message: None,
-                    },
-                    ExtensionInstallLocationStatus {
-                        enabled: Some(false),
-                        error: None,
-                        database: "db_where_it_is_currently_in_error_having_tried_to_enable_and_failed"
-                            .to_string(),
-                        schema: Some("public".to_string()),
-                        version: None,
-                        error_message: None,
-                    },
-                    ExtensionInstallLocationStatus {
-                        enabled: Some(false),
-                        error: None,
-                        database: "db_where_enable_failed".to_string(),
-                        schema: Some("public".to_string()),
-                        version: None,
-                        error_message: None,
-                    },
-                ],
-            },
-            ExtensionStatus {
-                name: "ext2".to_string(),
-                description: None,
-                locations: vec![
-                    ExtensionInstallLocationStatus {
-                        enabled: Some(true),
-                        error: None,
-                        database: "db2".to_string(),
-                        schema: Some("public".to_string()),
-                        version: None,
-                        error_message: None,
-                    },
-                    ExtensionInstallLocationStatus {
-                        enabled: Some(true),
-                        error: None,
-                        database: "db1".to_string(),
-                        schema: Some("public".to_string()),
-                        version: None,
-                        error_message: None,
-                    },
-                ],
-            },
-        ];
-
-        let result = determine_updated_extensions_status(&cdb, all_actually_installed_extensions);
-
-        // Update the extensions status
-        let cdb = CoreDB {
-            status: Some(CoreDBStatus {
-                extensions: Some(result),
-                ..CoreDBStatus::default()
-            }),
-            ..cdb
-        };
-
-        // Check that the current status is updated in the expected way from the provided actually_installed_extensions list
-        let location_status =
-            get_location_status(&cdb, "ext1", "db_where_its_available_and_disabled").unwrap();
-        assert_eq!(location_status.enabled, Some(false));
-        let location_status =
-            get_location_status(&cdb, "ext1", "db_where_its_available_and_enabled").unwrap();
-        assert_eq!(location_status.enabled, Some(true));
-        let location_status = get_location_status(
-            &cdb,
-            "ext1",
-            "db_where_its_available_and_disabled_missing_from_status",
-        )
-        .unwrap();
-        assert_eq!(location_status.enabled, Some(false));
-        let location_status = get_location_status(
-            &cdb,
-            "ext1",
-            "db_where_its_available_and_enabled_missing_from_status",
-        )
-        .unwrap();
-        assert_eq!(location_status.enabled, Some(true));
-        // Toggling and extension back to false, it should clear the error
-        let location_status = get_location_status(
-            &cdb,
-            "ext1",
-            "db_where_it_is_currently_in_error_having_tried_to_enable_and_failed",
-        )
-        .unwrap();
-        assert_eq!(location_status.enabled, Some(false));
-        assert!(!location_status.error.unwrap());
-        assert!(location_status.error_message.is_none());
-        // Toggling and extension back to false because missing, it should remove from status
-        assert!(get_location_status(
-            &cdb,
-            "ext1",
-            "db_where_it_is_currently_in_error_having_tried_to_enable_and_failed_because_missing",
-        )
-        .is_none());
-        // It should retain error message when it failed on a previous attempt
-        let location_status = get_location_status(&cdb, "ext1", "db_where_enable_failed").unwrap();
-        assert_eq!(location_status.enabled, Some(false));
-        assert!(location_status.error.unwrap());
-        assert!(location_status.error_message.is_some());
-        let location_status = get_location_status(&cdb, "ext1", "db_where_its_not_available").unwrap();
-        assert_eq!(location_status.enabled, None);
-        assert!(location_status.error.unwrap());
-        assert!(location_status.error_message.is_some());
-
-        let extension_locations_to_toggle = determine_extension_locations_to_toggle(&cdb);
-        // We just make this CDB so that we can use our getter function to
-        // search through the extension results from determine_extension_locations_to_toggle
-        let cdb_spec_check = CoreDB {
-            spec: CoreDBSpec {
-                extensions: extension_locations_to_toggle,
-                ..CoreDBSpec::default()
-            },
-            ..cdb
-        };
-
-        // When available and disabled, requesting to enable, we should try to toggle it
-        let location =
-            get_location_spec(&cdb_spec_check, "ext1", "db_where_its_available_and_disabled").unwrap();
-        assert!(location.enabled);
-        // When available and enabled, requesting to disable, we should try to toggle it
-        let location =
-            get_location_spec(&cdb_spec_check, "ext1", "db_where_its_available_and_enabled").unwrap();
-        assert!(!location.enabled);
-        // When available and disabled, requesting to enable, we should try to toggle it
-        let location = get_location_spec(
-            &cdb_spec_check,
-            "ext1",
-            "db_where_its_available_and_disabled_missing_from_status",
-        )
-        .unwrap();
-        assert!(location.enabled);
-        // When available and enabled, requesting to disable, we should try to toggle it
-        let location = get_location_spec(
-            &cdb_spec_check,
-            "ext1",
-            "db_where_its_available_and_enabled_missing_from_status",
-        )
-        .unwrap();
-        assert!(!location.enabled);
-
-        // If we toggled an extension to True, but it failed to enable
-        // and then we toggle it back to false, then it does not need a toggle
-        // because it's already in the desired state as disabled
-        let location = get_location_spec(
-            &cdb_spec_check,
-            "ext1",
-            "db_where_it_is_currently_in_error_having_tried_to_enable_and_failed",
-        );
-        assert!(location.is_none());
-        // If we toggled an extension to True, but it failed to enable because missing
-        // and then we toggle it back to false, then it does not need a toggle
-        // because it's already in the desired state as disabled
-        let location = get_location_spec(
-            &cdb_spec_check,
-            "ext1",
-            "db_where_it_is_currently_in_error_having_tried_to_enable_and_failed_because_missing",
-        );
-        assert!(location.is_none());
-        // If we request to enable an extension that is not installed,
-        // we should not try to toggle it
-        let location = get_location_spec(&cdb_spec_check, "ext1", "db_where_its_not_available");
-        assert!(location.is_none());
-        // If we request to enable an extension that has previously failed to enable,
-        // we should not try to toggle it again
-        let location = get_location_spec(&cdb_spec_check, "ext1", "db_where_enable_failed");
-        assert!(location.is_none());
     }
 }
